+++ jormungandr-lib/src/interfaces/block0_configuration/slots_duration.rs
@@ -18,7 +18,7 @@ impl SlotDuration {
     ///
     /// assert_eq!(SlotDuration::MINIMUM, SlotDuration::new(1).unwrap())
     /// ```
-    pub const MINIMUM: Self = SlotDuration(MINIMUM_SLOT_DURATION);
+    pub const MINIMUM: Self = Self(MINIMUM_SLOT_DURATION);
     /// maximum value for the slot duration
     ///
     /// ```
@@ -26,7 +26,7 @@ impl SlotDuration {
     ///
     /// assert_eq!(SlotDuration::MAXIMUM, SlotDuration::new(255).unwrap())
     /// ```
-    pub const MAXIMUM: Self = SlotDuration(MAXIMUM_SLOT_DURATION);
+    pub const MAXIMUM: Self = Self(MAXIMUM_SLOT_DURATION);
 
     /// create a new SlotDuration value
     ///
@@ -37,7 +37,7 @@ impl SlotDuration {
         if v < MINIMUM_SLOT_DURATION || MAXIMUM_SLOT_DURATION < v {
             None
         } else {
-            Some(SlotDuration(v))
+            Some(Self(v))
         }
     }
 }