+++ jormungandr-lib/src/interfaces/config_params.rs
@@ -2,7 +2,8 @@ use super::{
     ActiveSlotCoefficient, BlockContentMaxSize, CommitteeIdDef, ConsensusLeaderId,
     ConsensusVersionDef, DiscriminationDef, EpochStabilityDepth, FeesGoTo, KesUpdateSpeed,
     LinearFeeDef, NumberOfSlotsPerEpoch, PerCertificateFeeDef, PerVoteCertificateFeeDef,
-    PoolParticipationCapping, Ratio, RewardParams, SlotDuration, TaxType, Value,
+    PoolParticipationCapping, ProposalExpiration, Ratio, RewardParams, SlotDuration, TaxType,
+    Value,
 };
 use crate::time::SecondsSinceUnixEpoch;
 use chain_addr::Discrimination;
@@ -41,8 +42,7 @@ pub enum ConfigParam {
     RemoveBftLeader(ConsensusLeaderId),
     #[serde(with = "LinearFeeDef")]
     LinearFee(LinearFee),
-    // TODO implement interface
-    ProposalExpiration(u32),
+    ProposalExpiration(ProposalExpiration),
     KesUpdateSpeed(KesUpdateSpeed),
     TreasuryAdd(Value),
     TreasuryParams(TaxType),
@@ -69,6 +69,10 @@ pub enum FromConfigParamError {
     SlotDuration(#[from] super::block0_configuration::TryFromSlotDurationError),
     #[error("Invalid FeesGoTo setting")]
     FeesGoTo(#[from] super::block0_configuration::TryFromFeesGoToError),
+    #[error("Invalid active slot coefficient value")]
+    ActiveSlotCoefficient(#[from] super::block0_configuration::TryFromActiveSlotCoefficientError),
+    #[error("Invalid KES Update speed value")]
+    KesUpdateSpeed(#[from] super::block0_configuration::TryFromKesUpdateSpeedError),
 }
 
 impl From<ConfigParams> for ConfigParamsLib {
@@ -90,16 +94,13 @@ impl From<ConfigParam> for ConfigParamLib {
             ConfigParam::SlotsPerEpoch(val) => Self::SlotsPerEpoch(val.into()),
             ConfigParam::SlotDuration(val) => Self::SlotDuration(val.into()),
             ConfigParam::EpochStabilityDepth(val) => Self::EpochStabilityDepth(val.into()),
-            ConfigParam::ConsensusGenesisPraosActiveSlotsCoeff(val) => {
-                Self::ConsensusGenesisPraosActiveSlotsCoeff(val.0)
-            }
+            ConfigParam::ConsensusGenesisPraosActiveSlotsCoeff(val) => Self::from(val),
             ConfigParam::BlockContentMaxSize(val) => Self::BlockContentMaxSize(val.into()),
             ConfigParam::AddBftLeader(val) => Self::AddBftLeader(val.into()),
             ConfigParam::RemoveBftLeader(val) => Self::RemoveBftLeader(val.into()),
             ConfigParam::LinearFee(val) => Self::LinearFee(val),
-            // TODO implement
-            ConfigParam::ProposalExpiration(val) => Self::ProposalExpiration(val),
-            ConfigParam::KesUpdateSpeed(val) => Self::KesUpdateSpeed(val.0),
+            ConfigParam::ProposalExpiration(val) => Self::ProposalExpiration(val.into()),
+            ConfigParam::KesUpdateSpeed(val) => Self::from(val),
             ConfigParam::TreasuryAdd(val) => Self::TreasuryAdd(val.into()),
             ConfigParam::TreasuryParams(val) => Self::TreasuryParams(val.into()),
             ConfigParam::RewardPot(val) => Self::RewardPot(val.into()),
@@ -137,8 +138,10 @@ impl TryFrom<ConfigParamLib> for ConfigParam {
             ConfigParamLib::EpochStabilityDepth(val) => {
                 Self::EpochStabilityDepth(EpochStabilityDepth::from(val))
             }
-            ConfigParamLib::ConsensusGenesisPraosActiveSlotsCoeff(val) => {
-                Self::ConsensusGenesisPraosActiveSlotsCoeff(ActiveSlotCoefficient(val))
+            config @ ConfigParamLib::ConsensusGenesisPraosActiveSlotsCoeff(_) => {
+                Self::ConsensusGenesisPraosActiveSlotsCoeff(ActiveSlotCoefficient::try_from(
+                    config,
+                )?)
             }
             ConfigParamLib::BlockContentMaxSize(val) => Self::BlockContentMaxSize(val.into()),
             ConfigParamLib::AddBftLeader(val) => Self::AddBftLeader(ConsensusLeaderId::from(val)),
@@ -146,9 +149,10 @@ impl TryFrom<ConfigParamLib> for ConfigParam {
                 Self::RemoveBftLeader(ConsensusLeaderId::from(val))
             }
             ConfigParamLib::LinearFee(val) => Self::LinearFee(val),
-            // TODO implement
-            ConfigParamLib::ProposalExpiration(val) => Self::ProposalExpiration(val),
-            ConfigParamLib::KesUpdateSpeed(val) => Self::KesUpdateSpeed(KesUpdateSpeed(val)),
+            ConfigParamLib::ProposalExpiration(val) => Self::ProposalExpiration(val.into()),
+            config @ ConfigParamLib::KesUpdateSpeed(_) => {
+                Self::KesUpdateSpeed(KesUpdateSpeed::try_from(config)?)
+            }
             ConfigParamLib::TreasuryAdd(val) => Self::TreasuryAdd(Value::from(val)),
             ConfigParamLib::TreasuryParams(val) => Self::TreasuryParams(TaxType::from(val)),
             ConfigParamLib::RewardPot(val) => Self::RewardPot(Value::from(val)),
@@ -183,8 +187,39 @@ mod test {
 
     impl Arbitrary for ConfigParam {
         fn arbitrary<G: Gen>(g: &mut G) -> Self {
-            let config = ConfigParamLib::arbitrary(g);
-            Self::try_from(config).unwrap()
+            match u8::arbitrary(g) % 30 {
+                0 => Self::Block0Date(Arbitrary::arbitrary(g)),
+                1 => Self::Discrimination(Arbitrary::arbitrary(g)),
+                2 => Self::ConsensusVersion(Arbitrary::arbitrary(g)),
+                3 => Self::SlotsPerEpoch(Arbitrary::arbitrary(g)),
+                4 => Self::SlotDuration(Arbitrary::arbitrary(g)),
+                5 => Self::ConsensusGenesisPraosActiveSlotsCoeff(Arbitrary::arbitrary(g)),
+                6 => Self::BlockContentMaxSize(Arbitrary::arbitrary(g)),
+                7 => Self::AddBftLeader(Arbitrary::arbitrary(g)),
+                8 => Self::RemoveBftLeader(Arbitrary::arbitrary(g)),
+                9 => Self::LinearFee(Arbitrary::arbitrary(g)),
+                10 => Self::ProposalExpiration(Arbitrary::arbitrary(g)),
+                11 => Self::TreasuryAdd(Arbitrary::arbitrary(g)),
+                12 => Self::RewardPot(Arbitrary::arbitrary(g)),
+                13 => Self::RewardParams(Arbitrary::arbitrary(g)),
+                14 => Self::PerCertificateFees(Arbitrary::arbitrary(g)),
+                15 => Self::FeesInTreasury(Arbitrary::arbitrary(g)),
+                16 => Self::AddCommitteeId(Arbitrary::arbitrary(g)),
+                17 => Self::RemoveCommitteeId(Arbitrary::arbitrary(g)),
+                18 => Self::PerVoteCertificateFees(Arbitrary::arbitrary(g)),
+                19 => Self::RewardPot(Arbitrary::arbitrary(g)),
+                20 => Self::RewardParams(Arbitrary::arbitrary(g)),
+                21 => Self::RewardParams(Arbitrary::arbitrary(g)),
+                22 => Self::FeesInTreasury(Arbitrary::arbitrary(g)),
+                23 => Self::RewardLimitNone,
+                24 => Self::RewardLimitByAbsoluteStake(Arbitrary::arbitrary(g)),
+                25 => Self::PoolRewardParticipationCapping(Arbitrary::arbitrary(g)),
+                26 => Self::AddCommitteeId(Arbitrary::arbitrary(g)),
+                27 => Self::RemoveCommitteeId(Arbitrary::arbitrary(g)),
+                28 => Self::PerCertificateFees(Arbitrary::arbitrary(g)),
+                29 => Self::TransactionMaxExpiryEpochs(Arbitrary::arbitrary(g)),
+                _ => unreachable!(),
+            }
         }
     }