+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - fix incorrect keys bech32 HRP by always using the ones provided by the library
 - update REST API: add new endpoint AccountVotes (/api/v1/votes/plan/account-votes/{account_id})
+- add new setting proposal expiration into the initial config/genesis
 
 ## Release 0.13.0